added MyNFTCard and moved logic to it
