+++ components/MyNFTCard/index.ts
@@ -1,0 +1,1 @@
+export { default as MyNFTCard } from "./MyNFTCard";
